2.0.5 (2.1.0 Preview)
Release Notes
---------------
* [ALL CONVERSIONS NOW AVAILABLE AS ACTIONS AS WELL AS
DEVICES](https://github.com/Colorado4Wheeler/Device-Extensions/wiki/Acti
ons#conversion-action)!  You can also output the result to a variable,
the console log or to your speakers via text to speech
* Added new device [Virtual Color/Ambiant Hue Device
Group](https://github.com/Colorado4Wheeler/Device-Extensions/wiki/Extens
ion:-Color-Hue-Group) that allows you to group together Hue bulbs so
you can control them in tandem, including changing colors
* Added new device [Relay To Dimmer
Conversion](https://github.com/Colorado4Wheeler/Device-Extensions/wiki/R
elay-to-Dimmer-Conversion) that allows you to mimic dimmer commands
(i.e., brightness) on a relay device by calibrating the On/Off times to
a full cycle to give you percentages of "On".  This was created as a
result of my own need to have my curtains (which are on/off cycle) to
work with HomeKit and be able to open to a certain percentage even
though my curtains don't support that - now they do!
* Added new device [Filter
Sensor](https://github.com/Colorado4Wheeler/Device-Extensions/wiki/Filte
r-Sensor), mostly a direct need for HomeKit that allows you to have a
filter sensor come on when it's time to change a filter.  Can be based
on days/weeks/months or the run time of your HVAC if you want to use it
for your furnace filter
* Added new [action to direct speech to a specific Airfoil
speaker](https://github.com/Colorado4Wheeler/Device-Extensions/wiki/Acti
ons#extended-speech) or the system audio and then disconnect from the
speaker (if using Airfoil), with the ability to inject variables or
device states dynamically into the spoken phrase
* Added all remaining values and functions to the Thermostat Wrapper.
As of now you can only use device states for values and action groups
for functions but if there is a call to do more in the future this may
be added
* Added API hooks into HomeKit Bridge
* Added HomeKit Bridge integration into the recently created Thermostat
Wrapper, thus beginning the process of migrating all Wrapper and Alias
devices from Homebridge Buddy to this and other 3rd party applications
instead
* Added ability to execute an action group on Time to Elapsed Minutes
conversions if the elapsed minutes exceeds the provided threshold
* Added real-time updates for Conversion Extensions when using Convert
Date/Time to Elapsed Minutes, meaning that the system will update these
every 60 seconds from their source since in both of these cases the
source data may not change but our status needs to.  Note that this
will add a slight increase to the CPU usage, but probably not more than
0.1% to 0.3% depending on how many of these conversions you are doing
* Added plugin configuration for device address updates so that now a
Conversion Extension can show the type of conversion it is doing in the
address or than the device or variable it is doing it on (all previous
versions did the latter)
* Changed the order of devices in the XML so they appear in a more
logical order when selecting a device model
* Fixed bug where conversions on variables may not work, now all
variable conversions work properly
* Changed the About URL in the plugin menu to point to the Git Wiki
pages instead of the Indigo forums
* Removed Support Data Dump, Comprehensive Data Dump and Check for
Updates from the plugin menu and added the Advanced Plugin Actions in
its place to clean up the menu and to remove the version checker made
obsolete by the Plugin Store
* Fixed upgrade issue where it would always indicate on plugin startup
that it was upgrading from version 1.53, it will now only do that the
first time and then go away
* Cleaned up user interface for Conversion Extensions
* Cleaned up user interface for the action menu and moved all device
specific actions under devices
* Removed Conversion Extension checkbox to indicate the conversion is
for a device or variable and made it a combobox so that other
conversions can be easily integrated (i.e., action groups, schedules,
etc)
* Changed logic so that if you choose Always True or Always False as a
conversion method that the type will change to static and you will no
longer see a device, variable or state
* Always True and Always False conversion devices will now always have
an address of Static Value of True or Static Value of False in the UI
* Added utf8 encoding translations to some core library functions that
reference name, particularly the ones that may reference a non-plugin
object for thread debugging since certain latin characters over 128 may
cause errors and it is confusing if it doesn't come from a D/E device